requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
autocommand==2.2.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
cheroot==10.0.1
CherryPy==18.9.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-flex-fields==1.0.2
gevent==24.2.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflect==7.2.1
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
kombu==5.3.7
more-itertools==10.2.0
msgpack==1.0.8
portend==3.2.0
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tempora==5.5.1
Twisted==24.3.0
txaio==23.1.1
typeguard==4.3.0
typing_extensions==4.12.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zc.lockfile==3.0.post1
zope.event==5.0
zope.interface==6.4.post2
